+++ packages/core/src/types.ts
@@ -1328,23 +1328,6 @@ export interface IPdfService extends Service {
     convertPdfToText(pdfBuffer: Buffer): Promise<string>;
 }
 
-export interface IWebSearchService extends Service {
-    search(
-        query: string,
-        runtime: IAgentRuntime,
-        options?: SearchOptions,
-    ): Promise<SearchResponse>;
-}
-
-export interface SearchOptions {
-    limit?: number;
-    type?: "news" | "general";
-    includeAnswer?: boolean;
-    searchDepth?: "basic" | "advanced";
-    includeImages?: boolean;
-    days?: number; // 1 means current day, 2 means last 2 days
-}
-
 export interface IAwsS3Service extends Service {
     uploadFile(
         imagePath: string,
@@ -1411,28 +1394,6 @@ export interface ITeeLogService extends Service {
     ): Promise<boolean>;
 }
 
-export type SearchImage = {
-    url: string;
-    description?: string;
-};
-
-export type SearchResult = {
-    title: string;
-    url: string;
-    content: string;
-    rawContent?: string;
-    score: number;
-    publishedDate?: string;
-};
-
-export type SearchResponse = {
-    answer?: string;
-    query: string;
-    responseTime: number;
-    images: SearchImage[];
-    results: SearchResult[];
-};
-
 export enum ServiceType {
     IMAGE_DESCRIPTION = "image_description",
     TRANSCRIPTION = "transcription",
@@ -1448,7 +1409,6 @@ export enum ServiceType {
     IRYS = "irys",
     TEE_LOG = "tee_log",
     GOPLUS_SECURITY = "goplus_security",
-    WEB_SEARCH = "web_search",
 }
 
 export enum LoggingLevel {
