+++ packages/plugin-web-search/src/actions/webSearch.ts
@@ -1,14 +1,14 @@
-import { elizaLogger } from "@elizaos/core";
 import {
     Action,
     HandlerCallback,
     IAgentRuntime,
     Memory,
     State,
+    elizaLogger
 } from "@elizaos/core";
-import { generateWebSearch } from "@elizaos/core";
-import { SearchResult } from "@elizaos/core";
 import { encodingForModel, TiktokenModel } from "js-tiktoken";
+import { WebSearchService } from "../services/webSearchService";
+import { SearchResult } from "../types";
 
 const DEFAULT_MAX_WEB_SEARCH_TOKENS = 4000;
 const DEFAULT_MODEL_ENCODING = "gpt-3.5-turbo";
@@ -67,10 +67,10 @@ export const webSearch: Action = {
         const webSearchPrompt = message.content.text;
         elizaLogger.log("web search prompt received:", webSearchPrompt);
 
-        elizaLogger.log("Generating image with prompt:", webSearchPrompt);
-        const searchResponse = await generateWebSearch(
+        const webSearchService = new WebSearchService();
+        await webSearchService.initialize(runtime);
+        const searchResponse = await webSearchService.search(
             webSearchPrompt,
-            runtime
         );
 
         if (searchResponse && searchResponse.results.length) {
