+++ packages/plugin-web-search/src/index.ts
@@ -1,12 +1,15 @@
-import { webSearch } from "./actions";
+import { webSearch } from "./actions/webSearch";
 import { Plugin } from "@elizaos/core";
+import { WebSearchService } from "./services/webSearchService";
 
 export const webSearchPlugin: Plugin = {
     name: "webSearch",
-    description: "Search web",
+    description: "Search the web and get news",
     actions: [webSearch],
     evaluators: [],
     providers: [],
+    services: [new WebSearchService()],
+    clients: [],
 };
 
 export default webSearchPlugin;
